Prevent broken hide checkbox
If users clicked on the hide checkbox very often and fast it happened
that an error appeared and it was not possible to (un-)ignore a query
anymore unless users would reload the page.

Now the situation is a lot better. If you click often and fast it
is more unlikely that you see an error. Also, this behavior doesn't
break the hide checkbox.