feat: Webview: Add tab group on smaller screen sizes, make view expand horizontally
Rewrote Webview.svelte so that there is a tab group below the general area that contains the previously collapsible sections.
To ensure maximum display of information, more and more sections will leave the tab group and get their own columns,
where their content is on full display (with side-scrolling), as the screen gets wider.