feat: Add feedback concerning the status of the forwarding
For the purpose of not having the button look as if clicking it does nothing,
the following feedback was implemented:

While waiting for the post request to finish, the button will display a spinner.
If it was successful, the button will display a check symbol.
On failure the modal will close, giving a better view onto the main view where the error message will appear.