Don't show number of import errors next to imports
Otherwise the users thinks this number is the same as the number
of documents that were not imported because of an error but an
error doesn't automatically mean that a document wasn't imported
especially in a non-strict mode.

Also hide the legend when there is only one set of data.