feat: Move aggregator add form to modal
Old location below the list made it look like an afterthought and didn't quite fit in visually.
It also made the purpose of the aggregator viewer confusing, as it held both the function of a general overview and a create view at once.

With the modal the 'add' functionality is now separated from the overview.
The button is also placed in a prominent place right beside the view-title,
ensuring that no scrolling and searching is needed to find it, no matter how long the list is.