Delay requests when user selects dates
Otherwise, when the user uses the scroll wheel to choose the month/year of a date
and is not scrolling fast enough one request per month/year is made.

With this commit requests also are aborted if new ones are made. In the end these
changes lead to less load for the backend.