fix: Make QueryTable name and description columns use text-wrap
Previously, setting long titles or descriptions for queries would cause them to be displayed inline and cause the columns in question to expand to the width of the longest entry.
The table could therefore flow out of the page, even with bigger screen sizes.

This commit causes long names and descriptions to wrap if they get too long.